// Frontend Server (Angular2)/app/models/hotel/hotel.ts
export class Hotel {
  name: string;
  description: string;
  address: string;
  coords: number[];
  facilities: any[];
  languages: string[];
  region: string;
  reviews: any[];
  valuableInfo: any[];
}


/*
Copyright 2016 Google Inc. All Rights Reserved.
Use of this source code is governed by an MIT-style license that
can be found in the LICENSE file at http://angular.io/license
*/